requirements.txt
asgiref==3.5.1
Django==3.2.13
django-cleanup==6.0.0
django-debug-toolbar==3.4.0
django-mapbox-location-field==2.0.0
django-environ==0.8.1
Pillow==9.1.0
pytz==2022.1
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
djangorestframework==3.13.1